Updated Controllers

+++ backend/controllers/product.controller.js
@@ -49,6 +49,13 @@ export const createProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({
+            success: false,
+            message: "Invalid Product ID"
+        });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({
